greenDao open helper and App created and added to manifest

+++ app/src/main/java/com/example/android/maktab6/controller/ViewPagerActivity.java
@@ -29,7 +29,7 @@ public class ViewPagerActivity extends AppCompatActivity {
     private TabLayout mTabLayout;
     private FloatingActionButton mActionButton;
     private FragmentStatePagerAdapter mAdapter;
-    private int mUserId;
+    private Long mUserId;
 
     private UUID mTaskUUID;
