+++ app/src/main/java/com/example/android/maktab6/database/TaskCursorWrapper.java
@@ -22,7 +22,7 @@ public class TaskCursorWrapper extends CursorWrapper {
         String description = getString(getColumnIndex(DBSchema.TaskTable.TaskColumns.DESCRIPTION));
         Date date = new Date(getLong(getColumnIndex(DBSchema.TaskTable.TaskColumns.DATE)));
         boolean done = getInt(getColumnIndex(DBSchema.TaskTable.TaskColumns.DONE)) != 0;
-        int userId = getInt(getColumnIndex(DBSchema.TaskTable.TaskColumns.USER_ID));
+        Long userId = getLong(getColumnIndex(DBSchema.TaskTable.TaskColumns.USER_ID));
 
         Task task = new Task(uuid, userId);
         task.setTitle(title);
