+++ app/src/main/java/com/example/android/maktab6/model/LoginUser.java
@@ -1,5 +1,5 @@
 package com.example.android.maktab6.model;
 
 public class LoginUser {
-    public static int userLogin;
+    public static Long userLogin;
 }
